[REF] mail: convert discuss.upgrade service to plugin
In this commit, we rewrite the discuss.upgrade service as a plugin.
For legacy purposes, we keep the discuss.upgrade service (as a service).
We also adapt the codebase to reflect the changes on the service.
