+++ odoo/upgrade_code/owl3-migration.py
@@ -1101,6 +1101,7 @@ MISC_WHITELIST = {
 # serviceName: [PluginClass, ImportPath]
 SERVICES_MAPPING = {
     "currency": ['CurrencyPlugin', '@web/webclient/currency_plugin'],
+    "discuss.upgrade": ['DiscussUpgradePlugin', '@mail/core/common/upgrade/upgrade_plugin'],
     "notification": ['NotificationPlugin', '@web/core/notifications/notification_plugin'],
 }
